Fix to work in bash version 5
In previous Bash versions, one could use a `continue` command in the
root scope of a function, if that function was always called inside
a loop (`for`, `while`, etc.). In that case, the `continue` would be
applied in the loop that called the function.

It does not work in bash5 anymore, so I've moved the `continue`
outside of the function.

Thanks Yanmarshus Bachtiar for the bug report and the patch.