Fix quoting for --history (and add tests)
This was a regression when improving the printf commands. Now the
argument for `--history` that shows up when the program quits, is
correctly quoted with single quotes (double quotes would have issues
with shell special characters).

Also added a new test to make sure both the `--history` command line
and the textual description show as expected when the program quits.