Add tests/features.md to unit test features
This new test file is intended to be a tester and act as a permanent
record of how a specific txt2regex feature behaves.

This allows us to detect any functionality changes in the future, making
it easier to do code refactoring and fixing bugs.

Note that since it is a Markdown file, it also contains descriptions of
the features and the current problems they have.

The [clitest](https://github.com/aureliojargas/clitest) tester is used
to run the commands described in this file and check if the output is
correct.