getCharList(): Improve handling of '-'
It must not be moved if it is the first char.

Note that the 'number' regexes needed to be changed, because before
their `[-+]` list was inadvertently swapped to `[+-]` by the old code.
Now this does not occur anymore, but the original regex was changed to
keep the exact same output from the previous txt2regex versions.