New regex tester (runs in a Docker container)
Txt2regex needs to know regex-related information for each program it
supports. For example: the list of metacharacters, how to escape a
metacharacter to match it literally, availability of POSIX character
classes.

Instead of relying in documentation to get that information, there's a
new `tests/regex-tester.sh` script that calls the real programs with
specially crafted regexes and sample texts, verifying how those programs
behave in "real life".

To have a permanent record, the output of this script is also saved to
this repository. This way we can detect changes in behavior when a
program version is updated.

To avoid having to install specific software in the developer machine, a
Docker image is used to isolate all the necessary software and this
script is run inside that image (via `make test-regex`).

- Remove all the obsoleted files from the old tester:
  - test-suite/javascript.html
  - test-suite/procmail-re-test.sh
  - test-suite/result.txt
  - test-suite/test-suite.sh

- New `tests/regex-tester.sh` script

- New `tests/regex-tester.txt` script output record

- New `tests/Dockerfile` image with all the txt2regex-supported programs
  installed and ready to be used

- New make target: `test-regex` to run the tester and save to the output
  file

In this commit, the new regex tester is supporting all the programs that
the previous `test-suite.sh` script used to support. New programs will
be added in following commits.