regex-tester.sh: Always perform full matches (part 1)
It was a mistake not enforcing full matches since the start. Partial
matches are a problem when test_type=match.

This is just an intermediary step to make the .txt diff easier to see
if there are behavior changes (none detected on visual inspection).

The next commit will change all the actual regexes to be fully
anchored (explicit is better than implicit) and the .txt file should
not change in behavior (only the regexes will have the $ added)