Makefile: Use Docker Hub for the regex-tester image
Previously, the Docker image was built locally every time the targets
`test-regex` or `test-regex-shell` were called.

The docker cache makes this fast after the first build, but one drawback
is that some of the installed apt packages could have been updated, and
that will reflect in changes in `tests/regex-tester.txt`.

To have a truly reproducible environment, with fixed versions for all
programs, now the Docker image is built only once and then published
(manually) to Docker Hub.

This fixed environment is necessary to add `make test-regex` to the CI
(to be done in the next commit).

The idea is to update the image tag (currently `2020-05-09`) to the
current date, whenever a fresh build with updated program versions is
necessary in the future.