Create S2_* and ax_* regex-related arrays from plain text
Have all the regex-related data as plain text, to make it easier to read
and edit. Some magic code will convert this plain text to the previously
used `S2_*` and `ax_*` global arrays, so no code changes will be
required, everything will keep working as it was.

Note that now "-" is used to represent an "empty" value, where
previously the empty string was used.

A side benefit is that shfmt can now be applied to the whole source
code, because previously those arrays where always reformatted to an
ugly misaligned state. The Makefile was updated to check if the main
script is properly shfmt-formatted.

Another benefit is getting rid of the global shellcheck disable for
SC1003, that used to raise warnings in the old arrays. Now it can be
targeted to the one line that still triggers it.