ax_* arrays refactor: Use - instead of , as "false" marker
The hyphen is already used as empty/false for the item zero of the
array, so let's use it everywhere.

Now the comma is not special at all in txt2regex.

Note that `getHasPosix()` also needed to be changed (because it relied
in the comma) and was refactored to use `getMeta()`.

All those changes should bring no functionality changes.