Bump minimal Bash version from 2.05 to 3.0
This is required by the use of `<<<` to compose the `S2_*` and `ax_*`
arrays.

I could use here-docs `<<` instead, but there's another reason to bump
this version: the oldest supported version in the official Docker image
for Bash is 3.0.

So getting an old 2.05 and put automated tests for it would be a
deviation for the procedure already in place for all the other versions.

Version 2.05 is from 2001 and 3.0 is from 2004. We're still supporting a
very old version anyway.

Now every supported Bash version is properly tested in the CI.