Makefile: Remove tgz-related targets/variables
At the end, the FILES variable was listing every file in this repository
except `.travis.yml`, so the tarball was just a copy from the whole
repository.

That does not make sense, because one can just get the same files from
the Git history. Besides that, GitHub already provides `.tar.gz` and
`.zip` files for every tagged commit in the `Releases` tab.

For the next releases, only tagging the commit will be enough. Users
will continue to have the ability to download a tarball if needed,
directly from GitHub.

So there's no need for the tgz-related data in the Makefile, I won't
generate it anymore.