tests: Fix weird clitest + head + GitHub Actions bug
The previous code fails when running in GitHub Actions:

    [FAILED #39, line 384] txt2regex --foo | head -n 3 | sed '3 s/:.*//'
    @@ -1,3 +1,21 @@
    +./txt2regex.sh: line 115: printf: write error: Broken pipe
    +./txt2regex.sh: line 117: printf: write error: Broken pipe
    +./txt2regex.sh: line 119: printf: write error: Broken pipe
    +./txt2regex.sh: line 121: printf: write error: Broken pipe
    +./txt2regex.sh: line 122: printf: write error: Broken pipe
    +./txt2regex.sh: line 123: printf: write error: Broken pipe
     --foo: invalid option

     usage

Even when changing `head -n 3` to `sed 3q` it failed with the same
error.

When doing a single sed after the first pipe it worked fine *once*.
In the next run it failed again, with the same error.
I don't know what's happening, to be honest.

Trying with a temporary file, to avoid piping the first command.