Use quotes around $escape_metachar in escCharList
The expression uin="${uin/\\/$escape_metachar$escape_metachar}" has
different output on bash 5.2 then on earlier versions. Namely,
$escape_metachar is not treated as a string literal, so the escape
characters are applied which results in half as many escape chars in the
output:

  $ bash --version | head -1
  GNU bash, version 5.2.0(1)-rc2 (x86_64-pc-linux-gnu)
  $ uin='[\]'; escape_metachar='\\'; echo ${uin/\\/$escape_metachar$escape_metachar}
  [\\]

  vs.

  $ bash --version | head -1
  GNU bash, version 5.1.16(1)-release (x86_64-pc-linux-gnu)
  $ uin='[\]'; escape_metachar='\\'; echo ${uin/\\/$escape_metachar$escape_metachar}
  [\\\\]

To fix this, add quotes around $escape_metachar in this substitution:

  $ bash --version | head -1
  GNU bash, version 5.2.0(1)-rc2 (x86_64-pc-linux-gnu)
  $ uin='[\]'; escape_metachar='\\'; echo ${uin/\\/"$escape_metachar$escape_metachar"}
  [\\\\]

On bash versions 4.2 and earlier, the extra quotes will show up in the
output, so make the use of quotes conditional on the bash major version
number being 5 or higher.