Makefile: test-bash: Early exit in the first error
Errors in the previous bash versions were undetected, because this
target would exit as successful when the last tested bash version
was successful.

Now with the early exit, we make sure this target is aborted and
exit with error when the first tested bash version fails.

The downside is that we don't get all the fails in a single run, in
case more than one bash version fails. Well, we can't have
everything.