TODO: remove deprecated item: regex rules
Since https://github.com/aureliojargas/txt2regex/pull/6, the regex
rules are based on the results of actually running the programs in a
Docker container, using a special script.

So now it's mandatory to first add the new program to that container
and properly test it. The process became even more complex as before,
but now it's reliable and tested.